+++ lib/.byebug_history
@@ -1,0 +1,6 @@
+@prompt.board_size
+@prompt = Messenger::Prompt.new
+Messenger::Prompt.new
+Messenger::Prompt
+Messenger
+@prompt
