--- lib/helpers.rb
@@ -1,5 +1,0 @@
-module Helpers
-    def available_tiles(board)
-        board.flatten.select{|tile| tile.blank?}
-    end
-end
