+++ lib/tic_tac_toe/exceptions.rb
@@ -1,0 +1,3 @@
+module TicTacToe
+  class InvalidGameMode < StandardError; end
+end
