+++ lib/tic_tac_toe/helpers.rb
@@ -1,0 +1,7 @@
+module TicTacToe
+  module Helpers
+    def available_tiles(board)
+      board.flatten.select{|tile| tile.blank?}
+    end
+  end
+end
